Move gui code to topbar module
