+++ neothesia/src/scene/playing_scene/mod.rs
@@ -29,44 +29,6 @@ use toast_manager::ToastManager;
 mod animation;
 mod top_bar;
 
-mod icons {
-    pub fn cone_icon() -> &'static str {
-        "\u{F2D2}"
-    }
-
-    pub fn gear_icon() -> &'static str {
-        "\u{F3E5}"
-    }
-
-    pub fn gear_fill_icon() -> &'static str {
-        "\u{F3E2}"
-    }
-
-    pub fn repeat_icon() -> &'static str {
-        "\u{f130}"
-    }
-
-    pub fn play_icon() -> &'static str {
-        "\u{f4f4}"
-    }
-
-    pub fn pause_icon() -> &'static str {
-        "\u{f4c3}"
-    }
-
-    pub fn left_arrow_icon() -> &'static str {
-        "\u{f12f}"
-    }
-
-    pub fn minus_icon() -> &'static str {
-        "\u{F2EA}"
-    }
-
-    pub fn plus_icon() -> &'static str {
-        "\u{F4FE}"
-    }
-}
-
 const LAYER_BG: usize = 0;
 const LAYER_FG: usize = 1;
 
@@ -275,259 +237,11 @@ impl Scene for PlayingScene {
 
         self.update_glow(delta);
 
-        TopBar::update(self, ctx, delta);
+        TopBar::update(self, ctx);
 
         {
             use nuon::v2 as nuon;
             let ui = &mut self.nuon_ui;
-            let win_w = ctx.window_state.logical_size.width;
-
-            nuon::translate()
-                .y(self.top_bar.topbar_expand_animation.animate_bool(
-                    -75.0 + 5.0,
-                    0.0,
-                    ctx.frame_timestamp,
-                ))
-                .build(ui, |ui| {
-                    nuon::quad()
-                        .size(win_w, 30.0 + 45.0)
-                        .color([37, 35, 42])
-                        .build(ui);
-
-                    // Left
-                    if nuon::button()
-                        .size(30.0, 30.0)
-                        .border_radius([5.0; 4])
-                        .icon(icons::left_arrow_icon())
-                        .build(ui)
-                    {
-                        ctx.proxy
-                            .send_event(NeothesiaEvent::MainMenu(Some(self.player.song().clone())))
-                            .ok();
-                    }
-
-                    // Center
-                    nuon::translate()
-                        .x({
-                            let pill_w = 45.0 * 2.0;
-                            win_w / 2.0 - pill_w / 2.0
-                        })
-                        .y(5.0)
-                        .build(ui, |ui| {
-                            if nuon::button()
-                                .size(45.0, 20.0)
-                                .color([67, 67, 67])
-                                .hover_color([87, 87, 87])
-                                .preseed_color([97, 97, 97])
-                                .border_radius([10.0, 0.0, 10.0, 0.0])
-                                .icon(icons::minus_icon())
-                                .text_justify(nuon::TextJustify::Left)
-                                .build(ui)
-                            {
-                                ctx.config
-                                    .set_speed_multiplier(ctx.config.speed_multiplier() - 0.1);
-                            }
-
-                            nuon::label()
-                                .text(format!(
-                                    "{}%",
-                                    (ctx.config.speed_multiplier() * 100.0).round()
-                                ))
-                                .size(45.0 * 2.0, 20.0)
-                                .build(ui);
-
-                            if nuon::button()
-                                .size(45.0, 20.0)
-                                .x(45.0)
-                                .color([67, 67, 67])
-                                .hover_color([87, 87, 87])
-                                .preseed_color([97, 97, 97])
-                                .border_radius([0.0, 10.0, 0.0, 10.0])
-                                .icon(icons::plus_icon())
-                                .text_justify(nuon::TextJustify::Right)
-                                .build(ui)
-                            {
-                                ctx.config
-                                    .set_speed_multiplier(ctx.config.speed_multiplier() + 0.1);
-                            }
-                        });
-
-                    // Right
-                    nuon::translate()
-                        .x(ctx.window_state.logical_size.width - 30.0)
-                        .build(ui, |ui| {
-                            if nuon::button()
-                                .size(30.0, 30.0)
-                                .x(0.0)
-                                .border_radius([5.0; 4])
-                                .icon(if self.top_bar.settings_active {
-                                    icons::gear_fill_icon()
-                                } else {
-                                    icons::gear_icon()
-                                })
-                                .build(ui)
-                            {
-                                self.top_bar.settings_active = !self.top_bar.settings_active;
-                            }
-
-                            if nuon::button()
-                                .size(30.0, 30.0)
-                                .x(-30.0)
-                                .border_radius([5.0; 4])
-                                .icon(icons::repeat_icon())
-                                .build(ui)
-                            {
-                                self.top_bar.looper_active = !self.top_bar.looper_active;
-
-                                // Looper enabled for the first time
-                                if self.top_bar.looper_active
-                                    && self.top_bar.loop_start.is_zero()
-                                    && self.top_bar.loop_end.is_zero()
-                                {
-                                    self.top_bar.loop_start = self.player.time();
-                                    self.top_bar.loop_end =
-                                        self.player.time() + Duration::from_secs(5);
-                                }
-                            }
-
-                            if nuon::button()
-                                .size(30.0, 30.0)
-                                .x(-60.0)
-                                .border_radius([5.0; 4])
-                                .icon(if self.player.is_paused() {
-                                    icons::play_icon()
-                                } else {
-                                    icons::pause_icon()
-                                })
-                                .build(ui)
-                            {
-                                self.player.pause_resume();
-                            }
-                        });
-
-                    // ProggressBar
-                    nuon::translate().y(30.0).build(ui, |ui| {
-                        let h = 45.0;
-                        let w = ctx.window_state.logical_size.width;
-                        let progress_w = w * self.player.percentage();
-
-                        let loop_start = self.top_bar.loop_start;
-                        let loop_start = self.player.time_to_percentage(&loop_start) * w;
-
-                        let loop_end = self.top_bar.loop_end;
-                        let loop_end = self.player.time_to_percentage(&loop_end) * w;
-
-                        let loop_h = h + 10.0;
-
-                        let (loop_start_active, loop_end_active) = if self.top_bar.looper_active {
-                            let loop_start_ev = nuon::click_area("LooperStart")
-                                .x(loop_start)
-                                .width(5.0)
-                                .height(loop_h)
-                                .build(ui);
-                            if loop_start_ev.is_pressed() {
-                                let x = ctx.window_state.cursor_logical_position.x;
-                                let w = ctx.window_state.logical_size.width;
-                                let p = x / w;
-
-                                if p * w < loop_end - 10.0 {
-                                    self.top_bar.loop_start = self.player.percentage_to_time(p);
-                                }
-                            }
-
-                            let loop_end_ev = nuon::click_area("LooperEnd")
-                                .x(loop_end)
-                                .width(5.0)
-                                .height(loop_h)
-                                .build(ui);
-                            if loop_end_ev.is_pressed() {
-                                let x = ctx.window_state.cursor_logical_position.x;
-                                let w = ctx.window_state.logical_size.width;
-                                let p = x / w;
-
-                                if p * w > loop_start + 10.0 {
-                                    self.top_bar.loop_end = self.player.percentage_to_time(p);
-                                }
-                            }
-
-                            (
-                                loop_start_ev.is_hovered() || loop_start_ev.is_pressed(),
-                                loop_end_ev.is_hovered() || loop_end_ev.is_pressed(),
-                            )
-                        } else {
-                            (false, false)
-                        };
-
-                        match nuon::click_area("ProggressBar").size(w, h).build(ui) {
-                            nuon::ClickAreaEvent::PressStart => {
-                                if !self.rewind_controller.is_rewinding() {
-                                    self.rewind_controller.start_mouse_rewind(&mut self.player);
-
-                                    let x = ctx.window_state.cursor_logical_position.x;
-                                    let w = ctx.window_state.logical_size.width;
-
-                                    let p = x / w;
-                                    self.player.set_percentage_time(p);
-                                    self.keyboard.reset_notes();
-                                }
-                            }
-                            nuon::ClickAreaEvent::PressEnd { .. } => {
-                                self.rewind_controller.stop_rewind(&mut self.player);
-                            }
-                            nuon::ClickAreaEvent::Idle { .. } => {}
-                        }
-
-                        nuon::quad()
-                            .size(progress_w, h)
-                            .color([56, 145, 255])
-                            .build(ui);
-
-                        for m in self.player.song().file.measures.iter() {
-                            let length = self.player.length().as_secs_f32();
-                            let start = self.player.leed_in().as_secs_f32() / length;
-                            let measure = m.as_secs_f32() / length;
-
-                            let x = (start + measure) * w;
-
-                            let light_measure = nuon::Color::new(1.0, 1.0, 1.0, 0.5);
-                            let dark_measure = nuon::Color::new(0.4, 0.4, 0.4, 1.0);
-
-                            let color = if x < progress_w {
-                                light_measure
-                            } else {
-                                dark_measure
-                            };
-
-                            nuon::quad().x(x).size(1.0, h).color(color).build(ui);
-                        }
-
-                        if self.top_bar.looper_active {
-                            let color = [255, 56, 187];
-                            let white = [255; 3];
-
-                            nuon::quad()
-                                .x(loop_start)
-                                .width(loop_end - loop_start)
-                                .height(loop_h)
-                                .color([255, 56, 187, 90])
-                                .build(ui);
-
-                            nuon::quad()
-                                .x(loop_start)
-                                .width(5.0)
-                                .height(loop_h)
-                                .color(if loop_start_active { white } else { color })
-                                .build(ui);
-
-                            nuon::quad()
-                                .x(loop_end)
-                                .width(5.0)
-                                .height(loop_h)
-                                .color(if loop_end_active { white } else { color })
-                                .build(ui);
-                        }
-                    });
-                });
 
             for (rect, border_radius, color) in ui.quads.iter() {
                 self.quad_pipeline.push(
